// src/api/generated/models/propertyPurpose.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type PropertyPurpose = typeof PropertyPurpose[keyof typeof PropertyPurpose];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PropertyPurpose = {
  rent: 'rent',
  sale: 'sale',
  rent_or_sale: 'rent_or_sale',
} as const;