// src/api/generated/models/propertyStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type PropertyStatus = typeof PropertyStatus[keyof typeof PropertyStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PropertyStatus = {
  available: 'available',
  rented: 'rented',
  under_maintenance: 'under_maintenance',
  unavailable: 'unavailable',
  reserved: 'reserved',
  sold: 'sold',
} as const;